+++ Changelog.md
@@ -12,6 +12,7 @@
   - [#504](../../issues/504) - Allow to scope the treeview to an element (thanks @Laniusexcubitor)
   - [#508](../../issues/508) - Persist UseRegex in settings (thanks @corvinsz)
   - Improving clear filter handling and keyboard focus for property edits
+  - Showing sort indicators in the column headers of the property list
 
 ## 6.1.1